pyproject.toml
#-------------------------------------------------------------------------------
# Project: llama-kit
#-------------------------------------------------------------------------------

[project]

name = "llama-jax"
dynamic = ["version"]
requires-python = ">=3.12"

dependencies = [
    "blobfile>=3.0.0",
    "click>=8.1.7",
    "jax-metal>=0.1.1",
    "pandas>=2.2.3",
    "tiktoken>=0.8.0",
    "tqdm>=4.67.1",
]

[project.scripts]

llama-jax = "llama_jax.cli:cli"

[project.optional-dependencies]

cli = [
    "torch>=2.5.1",
]

#-------------------------------------------------------------------------------
# Build System
#-------------------------------------------------------------------------------

[build-system]

requires = ["hatchling"]
build-backend = "hatchling.build"

#-------------------------------------------------------------------------------
# Build Backend
#-------------------------------------------------------------------------------

[tool.hatch.version]

source = "env"
variable = "PY_VERSION"

#-------------------------------------------------------------------------------
# uv
#-------------------------------------------------------------------------------

[tool.uv]

dev-dependencies = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "jupyterlab>=4.3.2",
    "llama-models>=0.0.63",
    "matplotlib>=3.9.3",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.3",
    "python-dotenv>=1.0.1",
    "transformers>=4.47.1",
]
